pyproject.toml
[project]
name = "image-embedding-inference"
version = "0.0.1"
description = "Image Embedding Inference"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "gunicorn>=23.0.0",
    "pillow>=11.1.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.47.1",
    "uvicorn>=0.34.0",
]
